requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
pymysql==1.1.1
python-jose[cryptography]==3.5.0
pwdlib[argon2]==0.2.1
python-multipart==0.0.20
python-dotenv==1.1.1
aiosqlite==0.21.0
alembic==1.16.5
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1
cryptography==45.0.6
mysql-connector-python==9.4.0
sqlmodel==0.0.24